+++ src/declarations.d.ts
@@ -1,0 +1,4 @@
+declare module "@assets/*" {
+    const content: any;
+    export default content;
+}
