+++ src/screens/onboarding_1/components/BackButton.tsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { TouchableOpacity, StyleSheet } from "react-native";
+import { Ionicons } from "@expo/vector-icons";
+import { useNavigation } from "@react-navigation/native";
+
+export default function BackButton() {
+  const navigation = useNavigation();
+
+  return (
+    <TouchableOpacity 
+      style={styles.backButton} 
+      onPress={() => {
+        if (navigation.canGoBack()) {
+          navigation.goBack();
+        }
+        else{
+            alert('There is no back screen');
+        }
+      }}
+    >
+      <Ionicons name="arrow-back" size={24} color="white" />
+    </TouchableOpacity>
+  );
+}
+
+const styles = StyleSheet.create({
+  backButton: {
+    position: "absolute",
+    top: 50,
+    left: 20,
+    backgroundColor: "rgba(0,0,0,0.3)",
+    borderRadius: 20,
+    padding: 8,
+  },
+});
